Permissions base class and course/assignments endpoint secured (#222)
* Added base class and secured assignment endpoints

* [Untested] Secured most of course-endpoint via permissions

* Added possibility for group checking

* Rename function and isMapped optimization

+++ backend/src/main/kotlin/nl/utwente/horus/controllers/assignment/AssignmentSetsController.kt
@@ -1,14 +1,16 @@
 package nl.utwente.horus.controllers.assignment
 
-import nl.utwente.horus.exceptions.ForbiddenException
-import nl.utwente.horus.representations.assignment.AssignmentSetDtoFull
-import nl.utwente.horus.representations.assignment.AssignmentSetUpdateDto
 import nl.utwente.horus.auth.permissions.HorusPermission
+import nl.utwente.horus.auth.permissions.HorusPermissionType
 import nl.utwente.horus.auth.permissions.HorusResource
+import nl.utwente.horus.controllers.BaseController
+import nl.utwente.horus.entities.assignment.AssignmentSet
 import nl.utwente.horus.entities.group.Group
 import nl.utwente.horus.exceptions.InsufficientPermissionsException
 import nl.utwente.horus.representations.group.GroupDtoFull
 import nl.utwente.horus.representations.group.GroupDtoSummary
+import nl.utwente.horus.representations.assignment.AssignmentSetDtoFull
+import nl.utwente.horus.representations.assignment.AssignmentSetUpdateDto
 import nl.utwente.horus.services.assignment.AssignmentService
 import nl.utwente.horus.services.auth.HorusUserDetailService
 import org.springframework.beans.factory.annotation.Autowired
@@ -20,7 +22,7 @@ import org.springframework.web.bind.annotation.*
 @RestController
 @RequestMapping(path=["/api/assignmentSets"])
 @Transactional
-class AssignmentSetsController {
+class AssignmentSetsController: BaseController() {
 
     @Autowired
     lateinit var assignmentService: AssignmentService
@@ -30,22 +32,17 @@ class AssignmentSetsController {
 
     @GetMapping(path = ["/{assignmentSetId}"])
     fun getFullById(@PathVariable assignmentSetId: Long) : AssignmentSetDtoFull {
-        val assignmentSet = assignmentService.getAssignmentSetById(assignmentSetId)
-        val courseId = assignmentSet.course.id
+        verifyCoursePermission(AssignmentSet::class, assignmentSetId, HorusPermissionType.VIEW)
 
-        val hasAny = userDetailService.hasCoursePermission(courseId, HorusPermission.anyView(HorusResource.COURSE_ASSIGNMENTSET))
-        val hasOwn = userDetailService.hasCoursePermission(courseId,
-                HorusPermission.ownView(HorusResource.COURSE_ASSIGNMENTSET))
-        val isMappedToSet = assignmentService.isPersonMappedToAssignmentSet(userDetailService.getCurrentPerson(), assignmentSet)
-        if (!(hasAny || (hasOwn && isMappedToSet))) {
-            throw InsufficientPermissionsException()
-        }
+        val assignmentSet = assignmentService.getAssignmentSetById(assignmentSetId)
 
         return AssignmentSetDtoFull(assignmentSet)
     }
 
     @PutMapping(path = ["/{assignmentSetId}"])
     fun updateAssignmentSet(@PathVariable assignmentSetId: Long, @RequestBody dto: AssignmentSetUpdateDto) : AssignmentSetDtoFull {
+        verifyCoursePermission(AssignmentSet::class, assignmentSetId, HorusPermissionType.EDIT)
+
         val creator = userDetailService.getCurrentPerson()
         if (!userDetailService.hasCoursePermission(assignmentService.getAssignmentSetById(assignmentSetId).course.id,
                         HorusPermission.anyEdit(HorusResource.COURSE_ASSIGNMENTSET))) {
@@ -56,6 +53,7 @@ class AssignmentSetsController {
 
     @DeleteMapping(path = ["/{assignmentSetId}"])
     fun deleteAssignmentSet(@PathVariable assignmentSetId: Long) {
+        verifyCoursePermission(AssignmentSet::class, assignmentSetId, HorusPermissionType.DELETE)
         assignmentService.deleteAssignmentSet(assignmentService.getAssignmentSetById(assignmentSetId))
     }
